Change snapshot file format to contain request stream, just like xlog.
Snapshot file used to contain a stream of tuples, each
tuple prefixed with a small useless header. Change it
to contain a stream of REPLACE requests, so that
it can be loaded in just the same way as XLOG.
Change lsn in each request to be the lsn of the time the
snapshot is taken.

Disable (temporarily) snapshot.test, which tests compatibility
with the old snapshot.

@todo: add place for previous record's checksum into the row
header, to make sure row stream in snapshot/xlog files has no
gaps.